// freeboard_frontend_project/src/commons/types/generated/types.d.ts
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: any
}

export type IBoard = {
  __typename?: 'Board'
  _id: Scalars['ID']
  writer?: Maybe<Scalars['String']>
  title: Scalars['String']
  contents: Scalars['String']
  youtubeUrl: Scalars['String']
  likeCount: Scalars['Int']
  dislikeCount: Scalars['Int']
  boardAddress?: Maybe<IBoardAddress>
  user?: Maybe<IUser>
  createdAt: Scalars['DateTime']
  updatedAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
}

export type IBoardAddress = {
  __typename?: 'BoardAddress'
  _id: Scalars['ID']
  zipcode: Scalars['String']
  address: Scalars['String']
  addressDetail: Scalars['String']
  createdAt: Scalars['DateTime']
  updatedAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
}

export type IBoardComment = {
  __typename?: 'BoardComment'
  _id: Scalars['ID']
  writer?: Maybe<Scalars['String']>
  contents: Scalars['String']
  rating: Scalars['Float']
  user?: Maybe<IUser>
  createdAt: Scalars['DateTime']
  updatedAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
}
// ...
export type IPointTransaction = {
  __typename?: 'PointTransaction'
  _id: Scalars['ID']
  impUid: Scalars['ID']
  amount: Scalars['Int']
  balance: Scalars['Int']
  status: Scalars['String']
  statusDetail: Scalars['String']
  user?: Maybe<IUser>
  createdAt: Scalars['DateTime']
  updatedAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
}

export type IQuery = {
  __typename?: 'Query'
  fetchBoard: IBoard
  fetchBoards: Array<IBoard>
  fetchBoardsCount: Scalars['Int']
  fetchBoardsCountOfMine: Scalars['Int']
  fetchBoardsOfMine: Array<IBoard>
  fetchBoardsOfTheBest: Array<IBoard>
  fetchBoardComments: Array<IBoardComment>
  fetchPointTransactions: IPointTransaction
  fetchPointTransactionsCountOfBuying: IPointTransaction
  fetchPointTransactionsCountOfLoading: IPointTransaction
  fetchPointTransactionsCountOfSelling: IPointTransaction
  fetchPointTransactionsOfBuying: IPointTransaction
  fetchPointTransactionsOfLoading: IPointTransaction
  fetchPointTransactionsOfSelling: IPointTransaction
  fetchUseditem: IUseditem
  fetchUseditems: Array<IUseditem>
  fetchUseditemsCountIBought: Scalars['Int']
  fetchUseditemsCountIPicked: Scalars['Int']
  fetchUseditemsCountISold: Scalars['Int']
  fetchUseditemsIBought: Array<IUseditem>
  fetchUseditemsIPicked: Array<IUseditem>
  fetchUseditemsISold: Array<IUseditem>
  fetchUseditemsOfTheBest: Array<IUseditem>
  fetchUseditemQuestions: Array<IUseditemQuestion>
  fetchUseditemQuestionAnswers: Array<IUseditemQuestionAnswer>
  fetchUserLoggedIn: IUser
}

export type IQueryFetchBoardArgs = {
  boardId: Scalars['ID']
}
// ...
export type IUseditem = {
  __typename?: 'Useditem'
  _id: Scalars['ID']
  name: Scalars['String']
  remarks: Scalars['String']
  contents: Scalars['String']
  price: Scalars['Int']
  tags: Array<Scalars['String']>
  useditemAddress?: Maybe<IUseditemAddress>
  buyer: IUser
  seller: IUser
  soldAt?: Maybe<Scalars['DateTime']>
  createdAt: Scalars['DateTime']
  updatedAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
}

export type IUseditemAddress = {
  __typename?: 'UseditemAddress'
  id: Scalars['ID']
  zipcode: Scalars['String']
  address: Scalars['String']
  addressDetail: Scalars['String']
  lat: Scalars['Int']
  lng: Scalars['Int']
}

export type IUseditemQuestion = {
  __typename?: 'UseditemQuestion'
  _id: Scalars['ID']
  contents: Scalars['String']
  useditem: IUseditem
  user: IUser
  createdAt: Scalars['DateTime']
  updatedAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
}

export type IUseditemQuestionAnswer = {
  __typename?: 'UseditemQuestionAnswer'
  _id: Scalars['ID']
  contents: Scalars['String']
  useditemQuestion: IUseditemQuestion
  user: IUser
  createdAt: Scalars['DateTime']
  updatedAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
}

export type IUser = {
  __typename?: 'User'
  _id: Scalars['ID']
  email: Scalars['String']
  name: Scalars['String']
  userPoint?: Maybe<IUserPoint>
  createdAt: Scalars['DateTime']
  updatedAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
}

export type IUserPoint = {
  __typename?: 'UserPoint'
  _id: Scalars['ID']
  amount: Scalars['Int']
  user: IUser
  createdAt: Scalars['DateTime']
  updatedAt: Scalars['DateTime']
  deletedAt?: Maybe<Scalars['DateTime']>
}